Fix ResultOrderer component's file structure
